Move JSON logical tests out of dataclasses test script
